% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvErhard.R
\docType{package}
\name{pvErhard}
\alias{pvErhard}
\alias{pvErhard-package}
\title{ERHARD Plunger/Needle Valves (RKV) Sizing coefficients}
\format{
A Tibble with 39 elements
\describe{
\item{id}{Id of the valve}
\item{body}{Body type}
\item{typ}{type of control element}
\item{dn}{Diameter (in mm)}
\item{zvs}{Zeta Value full open}
\item{kvs}{calculated Flow coeficient full open (in m3/h)}
}
}
\source{
\url{https://www.erhard.de/en/}
}
\usage{
pvErhard
}
\description{
This package contains the Zeta values for the Erhard RKS valves

This package contains the Zeta values for the Erhard RKS valves
These data correspond to those published in Erhard's brochures.
}
\keyword{datasets}
